Styler listen med valg fra adressesøket
